tzcode93g.tar.gz and tzdata93g.tar.gz
The files
	tzcode93g.tar.gz
and
	tzdata93g.tar.gz
are now available for anonymous ftp from the ~ftp/pub directory on
elsie.nci.nih.gov (128.231.16.1).  Paul Eggert has provided a good
deal of historic information (based on Shanks), and there are some
code changes to deal with the buglets that crawled out in dealing with
the new information.

				--ado
Message-ID: <9311292018.AA11579@elsie.nci.nih.gov>