tzdata1997c.tar.gz
The file
	http://elsie.nci.nih.gov/pub/tzdata1997c.tar.gz
is now available; it incorporates the recently circulated changes in Israel.

I'm at work on the "zic -s" challenge. The cleanest way to deal with it seems
to be to ignore the "-s" flag when running through the rules and zones, then
eliminating "early" transition times when writing out the zone file; doing this
provides a way of figuring out what time type should be in effect at the
earliest instant to be represented in the file. Substantial regression work
will be needed, so stay tuned.

				--ado
Message-ID: <9703041511.AA02167@elsie.nci.nih.gov>