tzcode1998c.tar.gz and tzdata1998c.tar.gz
The files
	ftp://elsie.nci.nih.gov/pub/tzcode1998c.tar.gz
and
	ftp://elsie.nci.nih.gov/pub/tzcode1998c.tar.gz
are now available. These incorporate the proposed changes
circulated by Paul Eggert on the time zone mailing list a few
days ago (save the "French correction," on which I'll wait for
the dust to settle); also included are the symlink changes
circulated earlier this year and changes and additions to
Arts.htm.

				--ado
Message-ID: <9802281752.AA05228@elsie.nci.nih.gov>