data2010h
The file...
	ftp://elsie.nci.nih.gov/pub/tzdata2010h.tar.gz
...is now available; this reflects the changes circulated last week on the time zone mailing list.

I believe we're now update with this year's changes. There's still work to do on past changes in Ontario, Canada and Taipei.
There are also questions about just what happened on the Mexico side of the U.S.-Mexico border this year.
And there's the West Bank to be examined.

				--ado
Message-ID: <996D816825CFEA469870126E9050D3F0E010F5B4@NIHMLBX11.nih.gov>